fix tab switching bug
when on a PR, we can switch tabs (conversation, commits, checks and
files changed)
because GH uses a technology similar to turbolinks (where the whole
page isn't reloaded, only its content), izuna wasn't being run on tab
switch.

so we fix this by:
  - [x] listenning to the tab url changes
  - [x] using a cache to download the packages info only once every X seconds

Also, we used to send a dummy query from the content page to the
background page and wait for its response. Now we only send 1 message
from the background page to the content script!